requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
dj3-cloudinary-storage==0.0.6
Django==3.2.22
django-cloudinary-storage==0.3.0
django-filter==23.3
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
